Add files via upload
Update version of the Young Stellar Objects metrics. No additional software dependencies are required since the code automatically downloads the repository and the 3D extinction map.  The Dust Map name has been changed to avoid collisions with the default LSST MAF Dust map.  The metrics can be tested with a Notebook using the following commands:

import YoungStellarObjectsMetric
YoungStellarObjectsMetric.run_examples_datalab()

The notebook demonstrating the use of the metrics is  added to rubin_sim_notebooks. 
Please, contact  loredana.prisinzano@inaf.it for  updates  about the process of integration of this metrics into MAF and 
if you have any questions or need further info.